requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
bingmaps==0.3.7
celery==3.1.23
Django==1.10
djangorestframework==3.4.3
googlemaps==2.4.4
kombu==3.0.35
marshmallow==2.6.0
pytz==2016.6.1
requests==2.9.1
wheel==0.24.0
xmltodict==0.10.1
